+++ src/pages/admin/LibraryManage/Tab/BookManageTab.tsx
@@ -4,10 +4,12 @@ import { useGetBookManageListQuery, useDeleteBookMutation } from '@api/libraryMa
 import StandardTable from '@components/Table/StandardTable';
 import ActionButton from '@components/Button/ActionButton';
 import { Column, Row, ChildComponent } from '@components/Table/StandardTable.interface';
-import SearchSection from '@components/Section/SearchSection';
 import { IconButton } from '@mui/material';
 import { VscTrash } from 'react-icons/vsc';
+import { useSearchParams } from 'react-router-dom';
+import { BookListSearch } from '@api/dto';
 import AddBookModal from '../Modal/AddBookModal';
+import LibraryManageSearchSection from '../SearchSection/LibraryManageSearchSection';
 
 interface libraryManageRow {
   no: number;
@@ -34,16 +36,12 @@ const libraryManageColumn: Column<libraryManageRow>[] = [
 
 const BookManageTab = () => {
   const { page, getRowNumber } = usePagination();
-  const { data: bookManageListData } = useGetBookManageListQuery({ page });
-  const { mutate: deleteBookMutation } = useDeleteBookMutation();
+  const [searchParams] = useSearchParams();
+  const searchType = searchParams.get('searchType') as BookListSearch['searchType'];
+  const search = searchParams.get('search') as BookListSearch['search'];
 
-  const selectorList = [
-    { id: 'all', content: '도서명 + 저자' },
-    { id: 'title', content: '도서명' },
-    { id: 'author', content: '저자' },
-  ];
-  const [inputValue, setInputValue] = useState('');
-  const [selectorValue, setSelectorValue] = useState('all');
+  const { data: bookManageListData } = useGetBookManageListQuery({ page, searchType, search });
+  const { mutate: deleteBookMutation } = useDeleteBookMutation();
 
   const [addBookModalOpen, setAddBookModalOpen] = useState(false);
 
@@ -55,9 +53,7 @@ const BookManageTab = () => {
         return value;
     }
   };
-  const handleSearchButtonClick = () => {
-    // TODO 검색 API 호출
-  };
+
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const handleBookRowClick = ({ rowData }: { rowData: Row<libraryManageRow> }) => {
     // TODO 도서 수정 API 호출
@@ -71,14 +67,7 @@ const BookManageTab = () => {
   return (
     <>
       <div className="mb-5 flex justify-between space-x-2">
-        <SearchSection
-          options={selectorList}
-          selectorValue={selectorValue}
-          setSelectorValue={setSelectorValue}
-          inputValue={inputValue}
-          setInputValue={setInputValue}
-          onSearchButtonClick={handleSearchButtonClick}
-        />
+        <LibraryManageSearchSection />
         <ActionButton mode="add" onClick={() => setAddBookModalOpen(true)}>
           추가
         </ActionButton>
